+++ packages/manifest/src/index.ts
@@ -1,0 +1,3 @@
+export { StatsPlugin } from './StatsPlugin';
+export { ManifestManager } from './ManifestManager';
+export { StatsManager } from './StatsManager';