+++ apps/modernjs/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @module-federation/modernjs
 
+## 0.1.4
+
+### Patch Changes
+
+- @module-federation/enhanced@0.0.17
+
+## 0.1.3
+
+### Patch Changes
+
+- @module-federation/enhanced@0.0.16
+
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies [ba5bedd]
+  - @module-federation/enhanced@0.0.15
+
 ## 0.1.1
 
 ### Patch Changes